+++ src/main/java/com/lsy/www/service/UserService.java
@@ -21,7 +21,7 @@ public interface UserService {
      * @param id
      * @return 返回被影响的行数
      */
-    int deleteUserById(int id);
+    int deleteUserById(Long id);
 
     /**
      * 更新一个用户
@@ -35,7 +35,7 @@ public interface UserService {
      * @param id
      * @return
      */
-    UserEntity getUserById(int id);
+    UserEntity getUserById(Long id);
 
     /**
      * 根据条件查找多个用户
